feat: Add LLM Inference Service deployment component and refactor model template actions
