+++ src/lib/components/kind-viewer/kind-viewer-actions/modeltemplate/actions.svelte
@@ -4,14 +4,12 @@
 	import type { Schema } from '@sjsf/form';
 	import type { ValidateFunction } from 'ajv';
 
-	import Describe from '$lib/components/kind-viewer/kind-viewer-actions/default/describe.svelte';
 	import View from '$lib/components/kind-viewer/kind-viewer-actions/default/view.svelte';
 	import Button from '$lib/components/ui/button/button.svelte';
 	import * as DropdownMenu from '$lib/components/ui/dropdown-menu/index.js';
 
 	import Copy from './copy.svelte';
-	import DeployNaive from './deploy-naive.svelte';
-	import DeployWithLmCache from './deploy-with-lm-cache.svelte';
+	import Deploy from './deploy.svelte';
 
 	let {
 		cluster,
@@ -58,15 +56,24 @@
 			>
 				<View {schema} {object} />
 			</DropdownMenu.Item>
+			<DropdownMenu.Separator />
+			<DropdownMenu.Label>Manage</DropdownMenu.Label>
 			<DropdownMenu.Item
+				class="empty:hidden"
 				onSelect={(e) => {
 					e.preventDefault();
 				}}
 			>
-				<Describe {cluster} {namespace} {group} {version} {resource} {object} />
+				<Deploy
+					{cluster}
+					{namespace}
+					{schema}
+					{object}
+					onOpenChangeComplete={() => {
+						actionsOpen = false;
+					}}
+				/>
 			</DropdownMenu.Item>
-			<DropdownMenu.Separator />
-			<DropdownMenu.Label>Manage</DropdownMenu.Label>
 			<DropdownMenu.Item
 				onSelect={(e) => {
 					e.preventDefault();
@@ -74,42 +81,13 @@
 			>
 				<Copy
 					{cluster}
+					{namespace}
 					{group}
 					{version}
 					{kind}
 					{resource}
 					{schema}
 					{validate}
-					{object}
-					onOpenChangeComplete={() => {
-						actionsOpen = false;
-					}}
-				/>
-			</DropdownMenu.Item>
-			<DropdownMenu.Item
-				class="empty:hidden"
-				onSelect={(e) => {
-					e.preventDefault();
-				}}
-			>
-				<DeployNaive
-					{cluster}
-					{schema}
-					{object}
-					onOpenChangeComplete={() => {
-						actionsOpen = false;
-					}}
-				/>
-			</DropdownMenu.Item>
-			<DropdownMenu.Item
-				class="empty:hidden"
-				onSelect={(e) => {
-					e.preventDefault();
-				}}
-			>
-				<DeployWithLmCache
-					{cluster}
-					{schema}
 					{object}
 					onOpenChangeComplete={() => {
 						actionsOpen = false;
