--- src/lib/components/kind-viewer/kind-viewer-actions/modeltemplate/deploy-with-lm-cache.svelte
@@ -1,25 +1,0 @@
-<script lang="ts">
-	import type { ServingKserveIoV1Alpha2LLMInferenceServiceConfig } from '@otterscale/types';
-	import type { Schema } from '@sjsf/form';
-
-	let {
-		cluster,
-		schema,
-		object,
-		onOpenChangeComplete
-	}: {
-		cluster: string;
-		schema: Schema;
-		object: ServingKserveIoV1Alpha2LLMInferenceServiceConfig;
-		onOpenChangeComplete: () => void;
-	} = $props();
-
-	// svelte-ignore state_referenced_locally
-	void cluster;
-	// svelte-ignore state_referenced_locally
-	void schema;
-	// svelte-ignore state_referenced_locally
-	void object;
-	// svelte-ignore state_referenced_locally
-	void onOpenChangeComplete;
-</script>
