+++ src/lib/components/kind-viewer/kind-viewer-actions/modeltemplate/deploy.svelte
